[PUBLISHING] [CORRECTION] update word_count on correction
the "update_word_count" function is now in text_utils as it is a
generally useful function.

A new test is using cases handled by the word-count regex, the same test is done
in client.

SDFID-194
